+++ packages/mask/src/extension/popups/pages/Friends/common.tsx
@@ -32,3 +32,31 @@ export const PlatformIconMap: Record<SupportedPlatforms, GeneratedIcon> = {
     [NextIDPlatform.Unstoppable]: Icons.Unstoppable,
     [NextIDPlatform.Keybase]: Icons.Keybase,
 }
+
+export const UnsupportedPlatforms = [
+    NextIDPlatform.Bit,
+    NextIDPlatform.CyberConnect,
+    NextIDPlatform.REDDIT,
+    NextIDPlatform.SYBIL,
+    NextIDPlatform.EthLeaderboard,
+    NextIDPlatform.NextID,
+]
+
+export const PlatformSort: Record<NextIDPlatform, number> = {
+    [NextIDPlatform.Twitter]: 0,
+    [NextIDPlatform.GitHub]: 1,
+    [NextIDPlatform.Ethereum]: 2,
+    [NextIDPlatform.ENS]: 3,
+    [NextIDPlatform.LENS]: 4,
+    [NextIDPlatform.Keybase]: 5,
+    [NextIDPlatform.Farcaster]: 6,
+    [NextIDPlatform.SpaceId]: 7,
+    [NextIDPlatform.Unstoppable]: 8,
+    [NextIDPlatform.RSS3]: 9,
+    [NextIDPlatform.REDDIT]: 10,
+    [NextIDPlatform.SYBIL]: 11,
+    [NextIDPlatform.EthLeaderboard]: 12,
+    [NextIDPlatform.Bit]: 13,
+    [NextIDPlatform.CyberConnect]: 14,
+    [NextIDPlatform.NextID]: 15,
+}
